fix: Capability info lost on player death and cross dimension
