+++ src/main/java/com/sekwah/reskin/capabilities/ISkinData.java
@@ -1,6 +1,9 @@
 package com.sekwah.reskin.capabilities;
 
-public interface ISkinData {
+import net.minecraft.nbt.Tag;
+import net.minecraftforge.common.util.INBTSerializable;
+
+public interface ISkinData extends INBTSerializable<Tag> {
     String getSkin();
     void setSkin(String url);
     String getModelType();
